+++ vpin-studio-rest-client/src/main/java/de/mephisto/vpin/restclient/directb2s/BackglassServiceClient.java
@@ -3,11 +3,8 @@ package de.mephisto.vpin.restclient.directb2s;
 import de.mephisto.vpin.restclient.assets.AssetType;
 import de.mephisto.vpin.restclient.client.VPinStudioClient;
 import de.mephisto.vpin.restclient.client.VPinStudioClientService;
-import de.mephisto.vpin.restclient.directb2s.DirectB2SData;
-import de.mephisto.vpin.restclient.directb2s.DirectB2STableSettings;
-import de.mephisto.vpin.restclient.directb2s.DirectB2ServerSettings;
-import de.mephisto.vpin.restclient.jobs.JobExecutionResult;
 import de.mephisto.vpin.restclient.games.GameRepresentation;
+import de.mephisto.vpin.restclient.jobs.JobExecutionResult;
 import de.mephisto.vpin.restclient.util.FileUploadProgressListener;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -16,8 +13,6 @@ import org.springframework.http.ResponseEntity;
 
 import java.io.ByteArrayInputStream;
 import java.io.File;
-import java.net.URLEncoder;
-import java.nio.charset.StandardCharsets;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
@@ -46,28 +41,28 @@ public class BackglassServiceClient extends VPinStudioClientService {
     return getRestClient().get(API + "directb2s/" + gameId, DirectB2SData.class);
   }
 
-  public DirectB2SData getDirectB2SData(int emulatorId, String name) {
-    name = URLEncoder.encode(name, StandardCharsets.UTF_8).replace("+", "%20");
-    return getRestClient().get(API + "directb2s/" + emulatorId + "/" + name, DirectB2SData.class);
+  public DirectB2SData getDirectB2SData(DirectB2S directB2S) throws Exception {
+    return getRestClient().post(API + "directb2s/get", directB2S, DirectB2SData.class);
   }
 
-  public boolean deleteBackglass(int emulatorId, String name) {
-    name = URLEncoder.encode(name, StandardCharsets.UTF_8).replace("+", "%20");
-    return getRestClient().delete(API + "directb2s/" + emulatorId + "/" + name);
+  public boolean deleteBackglass(DirectB2S directB2S) throws Exception {
+    Map<String, Object> params = new HashMap<>();
+    params.put("fileName", directB2S.getFileName());
+    return getRestClient().post(API + "directb2s/delete", directB2S, Boolean.class);
   }
 
-  public boolean renameBackglass(int emulatorId, String name, String newName) throws Exception {
-    name = URLEncoder.encode(name, StandardCharsets.UTF_8).replace("+", "%20");
-    Map<String,Object> params = new HashMap<>();
+  public boolean renameBackglass(DirectB2S directB2S, String newName) throws Exception {
+    Map<String, Object> params = new HashMap<>();
     params.put("newName", newName);
-    return getRestClient().put(API + "directb2s/" + emulatorId + "/" + name, params, Boolean.class);
+    params.put("fileName", directB2S.getFileName());
+    return getRestClient().put(API + "directb2s", params, Boolean.class);
   }
 
-  public boolean duplicateBackglass(int emulatorId, String name) throws Exception {
-    name = URLEncoder.encode(name, StandardCharsets.UTF_8).replace("+", "%20");
-    Map<String,Object> params = new HashMap<>();
+  public boolean duplicateBackglass(DirectB2S directB2S) throws Exception {
+    Map<String, Object> params = new HashMap<>();
     params.put("duplicate", true);
-    return getRestClient().put(API + "directb2s/" + emulatorId + "/" + name, params, Boolean.class);
+    params.put("fileName", directB2S.getFileName());
+    return getRestClient().put(API + "directb2s", params, Boolean.class);
   }
 
   public List<DirectB2S> getBackglasses() {