+++ vpin-studio-rest-client/src/main/java/de/mephisto/vpin/restclient/games/GameListItem.java
@@ -2,7 +2,16 @@ package de.mephisto.vpin.restclient.games;
 
 public class GameListItem {
   private String name;
+  private String fileName;
   private int emuId;
+
+  public String getFileName() {
+    return fileName;
+  }
+
+  public void setFileName(String fileName) {
+    this.fileName = fileName;
+  }
 
   public String getName() {
     return name;